==> codes/3.1_mmp_mun_proj.Rmd <==
---
title:  'Migration - Projections'
author: "Facundo Danza & Eungik Lee"
---

# Pre-analysis

First, we clean the R-environment:
```{r}
rm(list = ls())
gc()
```

Then, we load (and install) the needed packages:
```{r}
if (!require("pacman")) install.packages("pacman")

pacman::p_load(data.table, tidyverse, janitor, readxl, xtable, scales, 
               future, future.apply, tictoc)
```

We then set the directories in which we have the data:
```{r}
dir = list()
dir$data         <- paste0(dirname(getwd()), "/data/")
dir$beta         <- paste0(dirname(getwd()), "/data/constructed/")
dir$commun       <- paste0(dirname(getwd()), "/data/commun/")
dir$weather      <- 
  paste0(dirname(getwd()), "/data/confidential/constructed/")
dir$constructed_data  <- 
  paste0(dirname(getwd()), "/data/confidential/constructed/")
dir$meteoblue    <- 
  paste0(dirname(getwd()), "/data/confidential/meteoblue/")
dir$mmp_final    <- 
  paste0(dirname(getwd()), "/data/confidential/constructed/")
dir$results      <- paste0(dirname(getwd()), "/tables/projection/")
dir$fig          <- paste0(dirname(getwd()), "/figures/climate_change/")
```


# Terraclimate - Weather Data

## Historical Data 

We start by opening the weather data on the municipalities for which some communities were surveyed in MMP (from the 2000 survey onward):
```{r}
mmp_mun_tmax = read.csv(paste0(dir$weather, 
                             "weather_mun_tmax_daymet.csv"))

mmp_mun_weather = mmp_mun_tmax

for (var in c("tmin", "prcp")) {
mmp_mun_round = read.csv(paste0(dir$weather, 
                             "weather_mun_",var,"_daymet.csv"))  

mmp_mun_weather = left_join(mmp_mun_weather, mmp_mun_round, 
                            by = c("state", "municipio", "year","month","n_obs"))
}

rm(list = c("mmp_mun_tmax", "mmp_mun_round"))
```


Since our data is yearly rather than monthly, we take the average of each variable, Furthermore, we only keep the months from April to September, the wet season in Mexico:
```{r}
month_start = 4
month_end   = 9

mmp_mun_weather_yr = mmp_mun_weather %>% 
  rename(tmax = tmax_avg, tmin = tmin_avg) %>%
  mutate(tavg = (tmax + tmin)/2) %>%
  filter(month >= month_start & month <= month_end) %>%
  group_by(state, municipio, year) %>%
  summarise(prcp_y_mean=mean(prcp_ttl),
            across(starts_with("t"), 
                   mean, .names = "{.col}_y_mean"))

```

As before, we need to normalize our dataset. We do so by taking the historical average over the period 1985-2014:
```{r}
start_yr_hist = 1985
end_yr_hist   = 2019

weather_norm = mmp_mun_weather_yr %>%
  filter(year <= end_yr_hist & year >= start_yr_hist) %>%
  group_by(municipio, state) %>%
  summarise(across(ends_with("y_mean"),
                   mean, .names = "{.col}_hist"), 
            across(ends_with("y_mean"),
                   sd, .names = "{.col}_sd_hist"))

```

```{r}
mmp_mun_weather_yr = mmp_mun_weather_yr %>% 
  right_join(weather_norm, by = c("municipio", "state"))

mmp_mun_weather_yr = 
  mmp_mun_weather_yr %>% 
  mutate(prcp_y_mean_dev = 
           (prcp_y_mean - prcp_y_mean_hist)/prcp_y_mean_sd_hist, 
         tavg_y_mean_dev = 
           (tavg_y_mean - tavg_y_mean_hist)/tavg_y_mean_sd_hist, 
         tmax_y_mean_dev = 
           (tmax_y_mean - tmax_y_mean_hist)/tmax_y_mean_sd_hist, 
         tmin_y_mean_dev = 
           (tmin_y_mean - tmin_y_mean_hist)/tmin_y_mean_sd_hist)
```


## Projections

### 2C
We load the weather projections:
```{r}
mmp_mun_tmax_2C = read.csv(paste0(dir$weather, 
                               "weather_mun_tmax_proj2C_1990.csv"))

years = 
  1991:2015

for (year_round in years) {
  mmp_mun_tmax_2C_round = 
    read.csv(paste0(dir$weather,
                    "weather_mun_tmax_proj2C_", 
                    year_round, ".csv"))
  
  mmp_mun_tmax_2C = 
    rbind(mmp_mun_tmax_2C, 
        mmp_mun_tmax_2C_round) 
}
```

```{r}
mmp_mun_tmin_2C = read.csv(paste0(dir$weather, 
                               "weather_mun_tmin_proj2C_1990.csv"))
years = 
  1991:2015

for (year_round in years) {
  mmp_mun_tmin_2C_round = 
    read.csv(paste0(dir$weather, 
                               "weather_mun_tmin_proj2C_", year_round, ".csv"))
  
  mmp_mun_tmin_2C = 
    rbind(mmp_mun_tmin_2C, 
        mmp_mun_tmin_2C_round) 
}
```

```{r}
mmp_mun_ppt_2C = read.csv(paste0(dir$weather, 
                               "weather_mun_ppt_proj2C_1990.csv"))

for (year_round in years) {
  mmp_mun_ppt_2C_round = 
    read.csv(paste0(dir$weather, 
                               "weather_mun_ppt_proj2C_", year_round, ".csv"))
  
  mmp_mun_ppt_2C= 
    rbind(mmp_mun_ppt_2C, 
        mmp_mun_ppt_2C_round) 
}
```


```{r}
mmp_mun_weather_2C = 
  mmp_mun_tmax_2C %>%
  left_join(mmp_mun_tmin_2C, 
            by = c("state", "municipio", "month", "year")) %>% 
  left_join(mmp_mun_ppt_2C, 
            by = c("state", "municipio", "month", "year"))

mmp_mun_weather_2C = 
  mmp_mun_weather_2C %>% 
  mutate(tavg = (tmax + tmin)/2)

rm(list = c("mmp_mun_tmax_2C", "mmp_mun_tmin_2C", "mmp_mun_ppt_2C"))
rm(list = ls(pattern = "_round"))
```

Since our data is yearly rather than monthly, we take the average of each variable, Furthermore, we only keep the months from April to September, the wet season for Mexican crops:
```{r}
month_start = 4
month_end   = 9

mmp_mun_weather_2C = mmp_mun_weather_2C %>% 
  filter(month >= month_start & month <= month_end) %>%
  group_by(state, municipio, year) %>%
  summarise(prcp_y_mean=mean(ppt),
           across(starts_with("t"),
                  mean, .names = "{.col}_y_mean")
)
```

We then proceed by taking the deviations for the historical mean of:

```{r}
mmp_mun_weather_2C = mmp_mun_weather_2C %>% 
  right_join(weather_norm, by = c("municipio", "state"))

mmp_mun_weather_2C = 
  mmp_mun_weather_2C %>% 
  mutate(prcp_y_mean_dev = 
           (prcp_y_mean - prcp_y_mean_hist)/prcp_y_mean_sd_hist, 
         tavg_y_mean_dev = 
           (tavg_y_mean - tavg_y_mean_hist)/tavg_y_mean_sd_hist, 
         tmax_y_mean_dev = 
           (tmax_y_mean - tmax_y_mean_hist)/tmax_y_mean_sd_hist, 
         tmin_y_mean_dev = 
           (tmin_y_mean - tmin_y_mean_hist)/tmin_y_mean_sd_hist)
```

```{r}
mmp_mun_weather_yr = 
  mmp_mun_weather_yr %>%
  mutate(across(ends_with("_dev"), ~ .x, 
                .names = "{.col}_true")) %>% 
  select(municipio, state, year, ends_with("_true"))
```

```{r}
mmp_mun_weather_all = 
  mmp_mun_weather_2C %>%
  left_join(mmp_mun_weather_yr, by = c("municipio", "state", "year"))
```

```{r}
tmax_harv_plt = 
  ggplot() +
  geom_density(data = mmp_mun_weather_all, 
               aes(x = tavg_y_mean_dev_true, fill = "Realized"), 
               alpha = 0.5) +
  geom_density(data = mmp_mun_weather_all, 
               aes(x = tavg_y_mean_dev, fill = "Projected"), 
               alpha = 0.5) +
  labs(x = "Z-score", 
       y = "Approximate Density", 
       fill = "Temperature: ") +
  scale_fill_manual(values = c("Realized" = "lightblue", 
                               "Projected" = "orange")) +
  theme_minimal() +
  theme(legend.position = "bottom")

tmax_harv_plt + 
  ggtitle("Maximum Temperature Distribution", 
         subtitle = "Climate Change")
```

```{r}
ggsave(paste0(dir$fig,"tmax_dev_2C_plt.png"), 
       width = 7, height = 4.5)
```


## Estimate effect

We load our estimates of the causal effect of weather on migration. Specifically, we load the beta_hat estimated in our main analysis:
```{r}
beta_hat = 
  read.csv(paste0(dir$beta,"beta_hat.csv"))
```

And its variance:
```{r}
beta_var = 
  read.csv(paste0(dir$beta,"beta_var.csv"))
```


And we simply project the immigration by multiply the effect of weather on migration over the expected change on weather:
```{r}
mmp_mun_weather_all = 
  mmp_mun_weather_all %>% ungroup() %>%
  mutate(imm_proj_prcp = (prcp_y_mean_dev - prcp_y_mean_dev_true)*beta_hat$beta_prcp,
         imm_proj_tmax = (tmax_y_mean_dev - tmax_y_mean_dev_true)*beta_hat$beta_tmax,
         imm_proj_tmin = (tmin_y_mean_dev - tmin_y_mean_dev_true)*beta_hat$beta_tmin)
```

```{r}
mmp_mun_weather_all %>%
  summarise(imm_proj_prcp = mean(imm_proj_prcp, na.rm = T), 
            imm_proj_tmax = mean(imm_proj_tmax, na.rm = T), 
            imm_proj_tmin = mean(imm_proj_tmin, na.rm = T))
```


And take the average of such result:
```{r}
table_2C_i = mmp_mun_weather_all %>%
  summarise("dev_proj_Precipitation (Z-score)"     = 
              mean(prcp_y_mean_dev - prcp_y_mean_dev_true, na.rm = T),
            "dev_proj_Max Temp (Z-score)" = 
              mean(tmax_y_mean_dev - tmax_y_mean_dev_true, na.rm = T), 
            "dev_proj_Min Temp (Z-score)" = 
              mean(tmin_y_mean_dev - tmin_y_mean_dev_true, na.rm = T)) %>%
  pivot_longer(cols = starts_with("dev_proj_"),
               names_to     = "Variable",
               values_to    = "Deviation",
               names_prefix = "dev_proj_")

table_2C_ii = mmp_mun_weather_all %>%
  summarise("imm_proj_Precipitation (Z-score)"         
            = mean(imm_proj_prcp, na.rm = T)*100,
            "imm_proj_Max Temp (Z-score)"     = 
              mean(imm_proj_tmin, na.rm = T)*100, 
            "imm_proj_Min Temp (Z-score)"     = 
              mean(imm_proj_tmax, na.rm = T)*100) %>%
  pivot_longer(cols = starts_with("imm_proj_"),
               names_to     = "Variable", 
               values_to    = "Migration (p.p)", 
               names_prefix = "imm_proj_")


table_2C = table_2C_i %>% 
  left_join(table_2C_ii, by = "Variable")
table_2C
```

```{r}
table_2C_i_SD = mmp_mun_weather_all %>%
  summarise("dev_proj_Precipitation (Z-score)"     = 
              sd(prcp_y_mean_dev - prcp_y_mean_dev_true, na.rm = T),
            "dev_proj_Max Temp (Z-score)" = 
              sd(tmax_y_mean_dev - tmax_y_mean_dev_true, na.rm = T), 
            "dev_proj_Min Temp (Z-score)" = 
              sd(tmin_y_mean_dev - tmin_y_mean_dev_true, na.rm = T)) %>%
  pivot_longer(cols = starts_with("dev_proj_"),
               names_to     = "Variable",
               values_to    = "Deviation",
               names_prefix = "dev_proj_")

table_2C_ii_SD = mmp_mun_weather_all %>%
  summarise("imm_proj_Precipitation (Z-score)"         
            = sd(imm_proj_prcp, na.rm = T)*100,
            "imm_proj_Max Temp (Z-score)"     = 
              sd(imm_proj_tmax, na.rm = T)*100, 
            "imm_proj_Min Temp (Z-score)"     = 
              sd(imm_proj_tmin, na.rm = T)*100) %>%
  pivot_longer(cols = starts_with("imm_proj_"),
               names_to     = "Variable", 
               values_to    = "Migration (p.p)", 
               names_prefix = "imm_proj_")

table_2C_SD = table_2C_i_SD %>% 
                       left_join(table_2C_ii_SD, by = "Variable") %>%
  mutate(across("Deviation":"Migration (p.p)", ~ paste0("(",round(.x,3),")")))

table_2C_SD
```

We add both results in one table:
```{r}
table_2C_string = table_2C %>% 
  mutate(across("Deviation":"Migration (p.p)", 
                ~ scales::comma(.x,accuracy = .01))) 

table_2C_all = 
  rbind(table_2C_string[1,], table_2C_SD[1,],
        table_2C_string[2,], table_2C_SD[2,],
        table_2C_string[3,], table_2C_SD[3,])

table_2C_all[c(2,4,6), 1] <- ""
table_2C_all
```

### 4C

We load the weather projections:
```{r}
mmp_mun_tmax_4C = read.csv(paste0(dir$weather, 
                                  "weather_mun_tmax_proj4C_1990.csv"))

years = 
  1991:2015

for (year_round in years) {
  mmp_mun_tmax_4C_round = 
    read.csv(paste0(dir$weather,
                    "weather_mun_tmax_proj4C_", 
                    year_round, ".csv"))
  
  mmp_mun_tmax_4C = 
    rbind(mmp_mun_tmax_4C, 
          mmp_mun_tmax_4C_round) 
}
```

```{r}
mmp_mun_tmin_4C = read.csv(paste0(dir$weather, 
                                  "weather_mun_tmin_proj4C_1990.csv"))
years = 
  1991:2015

for (year_round in years) {
  mmp_mun_tmin_4C_round = 
    read.csv(paste0(dir$weather, 
                    "weather_mun_tmin_proj4C_", year_round, ".csv"))
  
  mmp_mun_tmin_4C = 
    rbind(mmp_mun_tmin_4C, 
          mmp_mun_tmin_4C_round) 
}
```

```{r}
mmp_mun_ppt_4C = read.csv(paste0(dir$weather, 
                                 "weather_mun_ppt_proj4C_1990.csv"))

for (year_round in years) {
  mmp_mun_ppt_4C_round = 
    read.csv(paste0(dir$weather, 
                    "weather_mun_ppt_proj4C_", year_round, ".csv"))
  
  mmp_mun_ppt_4C= 
    rbind(mmp_mun_ppt_4C, 
          mmp_mun_ppt_4C_round) 
}
```


```{r}
mmp_mun_weather_4C = 
  mmp_mun_tmax_4C %>%
  left_join(mmp_mun_tmin_4C, 
            by = c("state", "municipio", "month", "year")) %>% 
  left_join(mmp_mun_ppt_4C, 
            by = c("state", "municipio", "month", "year"))

mmp_mun_weather_4C = 
  mmp_mun_weather_4C %>% 
  mutate(tavg = (tmax + tmin)/2)

rm(list = c("mmp_mun_tmax_4C", "mmp_mun_tmin_4C", "mmp_mun_ppt_4C"))
rm(list = ls(pattern = "_round"))
```

Since our data is yearly rather than monthly, we take the average of each variable, Furthermore, we only keep the months from April to September, the wet season for Mexican crops:
```{r}
month_start = 4
month_end   = 9

mmp_mun_weather_4C = mmp_mun_weather_4C %>% 
  filter(month >= month_start & month <= month_end) %>%
  group_by(state, municipio, year) %>%
  summarise(prcp_y_mean=mean(ppt),
            across(starts_with("t"),
                   mean, .names = "{.col}_y_mean")
  )
```

We then proceed by taking the deviations for the historical mean of:
  
```{r}
mmp_mun_weather_4C = mmp_mun_weather_4C %>% 
  right_join(weather_norm, by = c("municipio", "state"))

mmp_mun_weather_4C = 
  mmp_mun_weather_4C %>% 
  mutate(prcp_y_mean_dev = 
           (prcp_y_mean - prcp_y_mean_hist)/prcp_y_mean_sd_hist, 
         tavg_y_mean_dev = 
           (tavg_y_mean - tavg_y_mean_hist)/tavg_y_mean_sd_hist, 
         tmax_y_mean_dev = 
           (tmax_y_mean - tmax_y_mean_hist)/tmax_y_mean_sd_hist, 
         tmin_y_mean_dev = 
           (tmin_y_mean - tmin_y_mean_hist)/tmin_y_mean_sd_hist)
```

```{r}
mmp_mun_weather_yr = 
  mmp_mun_weather_yr %>%
  mutate(across(ends_with("_dev"), ~ .x, 
                .names = "{.col}_true")) %>% 
  select(municipio, state, year, ends_with("_true"))
```

```{r}
mmp_mun_weather_all = 
  mmp_mun_weather_4C %>%
  left_join(mmp_mun_weather_yr, by = c("municipio", "state", "year"))
```

```{r}
tmax_harv_plt = 
  ggplot() +
  geom_density(data = mmp_mun_weather_all, 
               aes(x = tavg_y_mean_dev_true, fill = "Realized"), 
               alpha = 0.5) +
  geom_density(data = mmp_mun_weather_all, 
               aes(x = tavg_y_mean_dev, fill = "Projected"), 
               alpha = 0.5) +
  labs(x = "Z-score", 
       y = "Approximate Density", 
       fill = "Temperature: ") +
  scale_fill_manual(values = c("Realized" = "lightblue", 
                               "Projected" = "orange")) +
  theme_minimal() +
  theme(legend.position = "bottom")

tmax_harv_plt + 
  ggtitle("Maximum Temperature Distribution", 
          subtitle = "Climate Change")
```

```{r}
ggsave(paste0(dir$fig,"tmax_dev_4C_plt.png"), 
       width = 7, height = 4.5)
```

## Estimate effect

We load our estimates of the causal effect of weather on migration. Specifically, we load the beta_hat estimated in our main analysis:
```{r}
beta_hat = 
  read.csv(paste0(dir$beta,"beta_hat.csv"))
```

And its variance:
```{r}
beta_var = 
  read.csv(paste0(dir$beta,"beta_var.csv"))
```


And we simply project the immigration by multiply the effect of weather on migration over the expected change on weather:
```{r}
mmp_mun_weather_all = 
  mmp_mun_weather_all %>% ungroup() %>%
  mutate(imm_proj_prcp = (prcp_y_mean_dev - prcp_y_mean_dev_true)*beta_hat$beta_prcp,
         imm_proj_tmax = (tmax_y_mean_dev - tmax_y_mean_dev_true)*beta_hat$beta_tmax,
         imm_proj_tmin = (tmin_y_mean_dev - tmin_y_mean_dev_true)*beta_hat$beta_tmin)
```

```{r}
mmp_mun_weather_all %>%
  summarise(imm_proj_prcp = mean(imm_proj_prcp, na.rm = T), 
            imm_proj_tmax = mean(imm_proj_tmax, na.rm = T), 
            imm_proj_tmin = mean(imm_proj_tmin, na.rm = T))
```


And take the average of such result:
```{r}
table_4C_i = mmp_mun_weather_all %>%
  summarise("dev_proj_Precipitation (Z-score)"     = 
              mean(prcp_y_mean_dev - prcp_y_mean_dev_true, na.rm = T),
            "dev_proj_Max Temp (Z-score)" = 
              mean(tmax_y_mean_dev - tmax_y_mean_dev_true, na.rm = T), 
            "dev_proj_Min Temp (Z-score)" = 
              mean(tmin_y_mean_dev - tmin_y_mean_dev_true, na.rm = T)) %>%
  pivot_longer(cols = starts_with("dev_proj_"),
               names_to     = "Variable",
               values_to    = "Deviation",
               names_prefix = "dev_proj_")

table_4C_ii = mmp_mun_weather_all %>%
  summarise("imm_proj_Precipitation (Z-score)"         
            = mean(imm_proj_prcp, na.rm = T)*100,
            "imm_proj_Max Temp (Z-score)"     = 
              mean(imm_proj_tmin, na.rm = T)*100, 
            "imm_proj_Min Temp (Z-score)"     = 
              mean(imm_proj_tmax, na.rm = T)*100) %>%
  pivot_longer(cols = starts_with("imm_proj_"),
               names_to     = "Variable", 
               values_to    = "Migration (p.p)", 
               names_prefix = "imm_proj_")


table_4C = table_4C_i %>% 
  left_join(table_4C_ii, by = "Variable")
table_4C
```

```{r}
table_4C_i_SD = mmp_mun_weather_all %>%
  summarise("dev_proj_Precipitation (Z-score)"     = 
              sd(prcp_y_mean_dev - prcp_y_mean_dev_true, na.rm = T),
            "dev_proj_Max Temp (Z-score)" = 
              sd(tmax_y_mean_dev - tmax_y_mean_dev_true, na.rm = T), 
            "dev_proj_Min Temp (Z-score)" = 
              sd(tmin_y_mean_dev - tmin_y_mean_dev_true, na.rm = T)) %>%
  pivot_longer(cols = starts_with("dev_proj_"),
               names_to     = "Variable",
               values_to    = "Deviation",
               names_prefix = "dev_proj_")

table_4C_ii_SD = mmp_mun_weather_all %>%
  summarise("imm_proj_Precipitation (Z-score)"         
            = sd(imm_proj_prcp, na.rm = T)*100,
            "imm_proj_Max Temp (Z-score)"     = 
              sd(imm_proj_tmax, na.rm = T)*100, 
            "imm_proj_Min Temp (Z-score)"     = 
              sd(imm_proj_tmin, na.rm = T)*100) %>%
  pivot_longer(cols = starts_with("imm_proj_"),
               names_to     = "Variable", 
               values_to    = "Migration (p.p)", 
               names_prefix = "imm_proj_")

table_4C_SD = table_4C_i_SD %>% 
  left_join(table_4C_ii_SD, by = "Variable") %>%
  mutate(across("Deviation":"Migration (p.p)", ~ paste0("(",round(.x,3),")")))

table_4C_SD
```

We add both results in one table:
```{r}
table_4C_string = table_4C %>% 
  mutate(across("Deviation":"Migration (p.p)", 
                ~ scales::comma(.x,accuracy = .01))) 

table_4C_all = 
  rbind(table_4C_string[1,], table_4C_SD[1,],
        table_4C_string[2,], table_4C_SD[2,],
        table_4C_string[3,], table_4C_SD[3,])

table_4C_all[c(2,4,6), 1] <- ""
table_4C_all
```


### Both

```{r}
tab_both =xtable(cbind(table_2C_all,table_4C_all[,2:3]), 
              caption = "Projected Illegal Migration",
              align   = c("c","l|","c","c|","c","c"),
              label   = "proj_imm")

print(tab_both,file=paste0(dir$results,"proj_imm.txt"),
      table.placement = "h",
      caption.placement="bottom", 
      include.rownames = FALSE, 
      include.colnames = TRUE, 
      hline.after = c(0, nrow(tab_both)), 
      add.to.row = list(pos = list(-1), command =c(" & \\multicolumn{2}{|c}{2ºC} & \\multicolumn{2}{|c}{4ºC} \\\\")))
```
